Added feature to use minimal signature for instrumentation functions
In order to save memory on embedded systems, this feature allows to
change the function signatures of the instrumentation functions to:

void __cyg_profile_func_enter(void *callsite);
void __cyg_profile_func_exit();

E.g. on ARM Cortex M cpu's the LR register gives the location where the
__cyg_profile function is called from, which is the address of the calling
function, plus a small offset. In some cases this is accurate enough to
replace the original first argument of __cyg_profile_func_enter().

The total overhead of enabling instrumentation functions is reduced with
around 60% using this flag, in terms of binary size.