emit warnings for unsigned > and <= comparisons against 0
Cases such as the following may very well be, or become, a bug:

    size_t space;
    ...
    while ( space > 0 )
    {
        ...
        space -= used;
    }

The issue is that, if 'space' never exactly hits 0, the loop will not stop. The bug in this code is that the type of space is unsigned.

By warning that the comparison actually boils down to space != 0, it becomes clear that something is wrong.